+++ Integrations/SplunkPy/SplunkPy.py
@@ -410,10 +410,14 @@ def splunk_job_create_command(service):
 
 def splunk_results_command(service):
     res = []
+    sid = demisto.args().get('sid', '')
     try:
-        job = service.job(demisto.args().get('sid', ''))
+        job = service.job(sid)
     except HTTPError as error:
-        return_error(error.message, error)
+        if error.message == 'HTTP 404 Not Found -- Unknown sid.':
+            demisto.results("Found no job for sid: {}".format(sid))
+        else:
+            return_error(error.message, error)
     else:
         for result in results.ResultsReader(job.results()):
             if isinstance(result, results.Message):
@@ -422,10 +426,7 @@ def splunk_results_command(service):
                 # Normal events are returned as dicts
                 res.append(result)
 
-        if not res:
-            demisto.results("Found no job for sid: " + demisto.args()['sid'])
-        else:
-            demisto.results({"Type": 1, "ContentsFormat": "json", "Contents": json.dumps(res)})
+        demisto.results({"Type": 1, "ContentsFormat": "json", "Contents": json.dumps(res)})
 
 
 def fetch_incidents(service):
@@ -604,6 +605,32 @@ def splunk_edit_notable_event_command(proxy):
     demisto.results('Splunk ES Notable events: ' + response_info['message'])
 
 
+def splunk_job_status(service):
+    sid = demisto.args().get('sid')
+    try:
+        job = service.job(sid)
+    except HTTPError as error:
+        if error.message == 'HTTP 404 Not Found -- Unknown sid.':
+            demisto.results("Not found job for SID: {}".format(sid))
+        else:
+            return_error(error.message, error)
+    else:
+        status = job.state.content.get('dispatchState')
+        entry_context = {
+            'SID': sid,
+            'Status': status
+        }
+        context = {'Splunk.JobStatus(val.SID && val.SID === obj.SID)': entry_context}
+        human_readable = tableToMarkdown('Splunk Job Status', entry_context)
+        demisto.results({
+            "Type": entryTypes['note'],
+            "Contents": entry_context,
+            "ContentsFormat": formats["json"],
+            "EntryContext": context,
+            "HumanReadable": human_readable
+        })
+
+
 def splunk_parse_raw_command():
     raw = demisto.args().get('raw', '')
     rawDict = rawToDict(raw)
@@ -688,6 +715,8 @@ def main():
         splunk_parse_raw_command()
     if demisto.command() == 'splunk-submit-event-hec':
         splunk_submit_event_hec_command()
+    if demisto.command() == 'splunk-job-status':
+        splunk_job_status(service)
 
 
 if __name__ in ['__main__', '__builtin__', 'builtins']: