+++ Packs/Base/Scripts/CommonServerPython/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+- Added support for successful empty responses (status code 204) in the base client
 
 
 ## [20.3.3] - 2020-03-18